Split List.hpp, moving Add, GetRange, etc to ListOperations.hpp; Tidy of Add operation;
