+++ server/include/fc/ListHandler.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <fc/FlatBuffers.hpp>
-#include <fc/List.hpp>
+#include <fc/ListOperations.hpp>
 #include <plog/Log.h>
 #include <map>
 #include <optional>
